+++ restapi/package.json
@@ -9,11 +9,12 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "body-parser": "^1.19.1",
+    "body-parser": "^1.19.2",
     "cors": "^2.8.5",
-    "express": "^4.17.2",
+    "express": "^4.17.3",
     "express-prom-bundle": "^6.4.1",
     "express-validator": "^6.14.0",
+    "mongoose": "^6.2.4",
     "prom-client": "^14.0.1"
   },
   "devDependencies": {
