+++ webapp/package.json
@@ -53,9 +53,9 @@
     "jest-cucumber": "^3.0.1",
     "jest-puppeteer": "^6.0.3",
     "puppeteer": "^13.0.1",
+    "react-scripts": "5.0.0",
     "start-server-and-test": "^1.14.0",
     "ts-jest": "^27.1.2",
-    "ts-node-dev": "^1.1.8",
-    "react-scripts": "5.0.0"
+    "ts-node-dev": "^1.1.8"
   }
 }
